chore: split custom color overrides into color_custom.go
color.go was 516 lines and mixed two concerns: built-in terminal
detection / ANSI escapes / cached YAML palettes, and a separate
user-customization layer that lets callers swap in hex colors via
ParseColor + CustomColorPalette.

Move the user-customization layer into a new color_custom.go (paired
with the existing color_custom_test.go):

  - ColorRole type + constants
  - CustomColor struct
  - CustomColorPalette type and methods (colorForRole, ColorCode,
    EntryPalette)
  - DefaultCustomColorPalette + defaultPalette global
  - resolvedPalette helper
  - ParseColor + ansiColorRef + ansiColorRefs + nearestANSI +
    colorDistSq

color.go drops the now-unused strconv import and is down to ~310
lines of "what diffyml ships by default". No behaviour change; pure
file reorganization.
